alldata/bblab_site/pyproject.toml
[project]
name = "bblab_site"
version = "0.0.0"
description = "BBLab web-site"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "AGPL-3.0" }
authors = [
    { name = "British Columbia Centre for Excellence in HIV/AIDS", email = "vmysak@bccfe.ca" },
]
classifiers = [
    "Framework :: Django",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
  "asgiref==3.8.1",
  "attrs==25.1.0",
  "bleach==6.2.0",
  "blinker==1.7.0",
  "cryptography==41.0.7",
  "dbus-python==1.3.2",
  "distro==1.9.0",
  "Django==5.1.5",
  "django-appconf==1.0.6",
  "django-axes==7.0.1",
  "django-classy-tags==4.1.0",
  "django-ipware==7.0.1",
  "django-js-asset==3.0.1",
  "django-mptt==0.16.0",
  "django-nyt==1.4.1",
  "django-sekizai==4.1.0",
  "drawsvg==2.4.0",
  "genetracks @ git+https://github.com/cfe-lab/genetracks.git@a7f20f644ab86b451e870c9ace0621af8ecae0df",
  "html5lib==1.1",
  "httplib2==0.20.4",
  "importlib_metadata==8.6.1",
  "iniconfig==2.0.0",
  "launchpadlib==1.11.0",
  "lazr.restfulclient==0.14.6",
  "lazr.uri==1.0.6",
  "Markdown==3.6",
  "mod-wsgi==5.0.2",
  "more-itertools==10.6.0",
  "mysqlclient==2.2.7",
  "numpy==2.2.2",
  "oauthlib==3.2.2",
  "packaging==24.2",
  "pillow==11.1.0",
  "pluggy==1.5.0",
  "py==1.11.0",
  "PyGObject==3.48.2",
  "PyJWT==2.7.0",
  "pymdown-extensions==10.5",
  "pyparsing==3.1.1",
  "pytest==8.3.4",
  "python-ipware==3.0.0",
  "pytz==2025.1",
  "PyYAML==6.0.2",
  "scipy==1.15.1",
  "setuptools==68.1.2",
  "six==1.16.0",
  "sorl-thumbnail==12.11.0",
  "sqlparse==0.5.3",
  "tinycss2==1.4.0",
  "wadllib==1.3.6",
  "wcwidth==0.2.13",
  "webencodings==0.5.1",
  "wheel==0.42.0",
  "wiki==0.11.2",
  "zipp==3.21.0",
]

[project.optional-dependencies]
test = [
  "pytest==8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "README.md",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = [
    ".",
]
# Exclude vendor/dependency directories and submodule repos that have their own test suites
norecursedirs = [
    "depend",
    ".venv",
    "build",
    "dist",
    "*.egg",
    ".git",
    "git",
    ".tox",
    "repo",  # Excludes submodule repos like tools/phage_i_expanded/repo/
    "blind_dating",
    "phylodating",
    "translate_DNA",
    "tests/phylodating",  # Exclude the test submodule copy
    "tools",  # Skip the empty Django test stubs
]
# Only collect test files from our actual project, not vendored dependencies
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
